document-service/requirements.txt
fastapi
uvicorn
httpx
python-pptx
python-docx
openpyxl
pytesseract
Pillow
pymupdf
